+++ scripts/transcode-to-json.py
@@ -57,8 +57,7 @@ if __name__ == "__main__":
 
     args = argparser.parse_args()
     try:
-        # open() does not work with Paths in Python 3.5
-        with codecs_open(str(args.input.resolve()), "r", "UTF-8") as f:
+        with codecs_open(args.input.resolve(), "r", "UTF-8") as f:
             data = f.read()
     except Exception:
         logger.error("Could not even read file: %s", args.input, exc_info=True)
@@ -92,8 +91,7 @@ if __name__ == "__main__":
 
     if args.output is not None:
         try:
-            # open() does not work with Paths in Python 3.5
-            with codecs_open(str(args.output.resolve()), "w", "UTF-8") as f:
+            with codecs_open(args.output.resolve(), "w", "UTF-8") as f:
                 f.write(data)
         except Exception:
             logger.error("Could open output file: %s", args.output, exc_info=True)